requirements.txt
torch
torchvision
streamlit
tiktoken
pillow
numpy
gdown